+++ Runtime/Scripts/Export/GltfUnlitMaterialExporter.cs
@@ -13,7 +13,7 @@ using Material = UnityEngine.Material;
 namespace GLTFast.Export
 {
     /// <summary>
-    /// Converts Unity Materials that use a glTFast shader to glTF materials
+    /// Converts Unity Materials that use the glTFast shader `glTF/Unlit` to glTF materials
     /// </summary>
     public class GltfUnlitMaterialExporter : IMaterialExport
     {
@@ -81,13 +81,7 @@ namespace GLTFast.Export
 
             if (GltfMaterialExporter.TryGetValue(unityMaterial, MaterialProperty.BaseColor, out Color baseColor))
             {
-                material.pbrMetallicRoughness.baseColorFactor = new[]
-                {
-                    baseColor.r,
-                    baseColor.g,
-                    baseColor.b,
-                    baseColor.a
-                };
+                material.pbrMetallicRoughness.BaseColor = baseColor.linear;
             }
 
             return material;