+++ Runtime/Scripts/GltfImport.cs
@@ -21,7 +21,7 @@ using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using System.Threading;
 using System;
-
+using System.Text;
 using GLTFast.Addons;
 using GLTFast.Jobs;
 #if MEASURE_TIMINGS
@@ -157,6 +157,7 @@ namespace GLTFast
 #endif
             ExtensionName.MaterialsPbrSpecularGlossiness,
             ExtensionName.MaterialsUnlit,
+            ExtensionName.MaterialsVariants,
             ExtensionName.TextureTransform,
             ExtensionName.MeshQuantization,
             ExtensionName.MaterialsTransmission,
@@ -470,17 +471,43 @@ namespace GLTFast
             CancellationToken cancellationToken = default
             )
         {
-            var firstBytes = new byte[4];
-
 #if UNITY_2021_3_OR_NEWER && NET_STANDARD_2_1
             await using
 #endif
             var fs = new FileStream(localPath, FileMode.Open, FileAccess.Read);
-            var bytesRead = await fs.ReadAsync(firstBytes, 0, firstBytes.Length, cancellationToken);
+            var result = await LoadStream(fs, uri, importSettings, cancellationToken);
+            fs.Dispose();
+            return result;
+        }
 
-            if (bytesRead != firstBytes.Length)
+        /// <summary>
+        /// Load glTF from a stream.
+        /// </summary>
+        /// <param name="stream">Stream of the glTF or glTF-Binary</param>
+        /// <param name="uri">Base URI for relative paths of external buffers or images</param>
+        /// <param name="importSettings">Import Settings (<see cref="ImportSettings"/> for details)</param>
+        /// <param name="cancellationToken">Token to submit cancellation requests. The default value is None.</param>
+        /// <returns>True if loading was successful, false otherwise</returns>
+        public async Task<bool> LoadStream(
+            Stream stream,
+            Uri uri = null,
+            ImportSettings importSettings = null,
+            CancellationToken cancellationToken = default)
+        {
+            if (!stream.CanRead)
             {
-                m_Logger?.Error(LogCode.Download, "Failed reading first bytes", localPath);
+                m_Logger?.Error(LogCode.StreamError, "Not readable");
+                return false;
+            }
+
+            var initialStreamPosition = stream.CanSeek
+                ? stream.Position
+                : -1L;
+
+            var firstBytes = new byte[4];
+            if (!await stream.ReadToArrayAsync(firstBytes, 0, firstBytes.Length, cancellationToken))
+            {
+                m_Logger?.Error(LogCode.StreamError, "First bytes");
                 return false;
             }
 
@@ -488,32 +515,52 @@ namespace GLTFast
 
             if (GltfGlobals.IsGltfBinary(firstBytes))
             {
-                var data = new byte[fs.Length];
-                for (var i = 0; i < firstBytes.Length; i++)
+                // Read the rest of the header
+                var glbHeader = new byte[8];
+                if (!await stream.ReadToArrayAsync(glbHeader, 0, glbHeader.Length, cancellationToken))
                 {
-                    data[i] = firstBytes[i];
-                }
-                var length = (int)fs.Length - 4;
-                var read = await fs.ReadAsync(data, 4, length, cancellationToken);
-                fs.Close();
-                if (read != length)
-                {
-                    m_Logger?.Error(LogCode.Download, "Failed reading data", localPath);
+                    m_Logger?.Error(LogCode.StreamError, "glb header");
                     return false;
+                }
+                // Length of the entire glTF, including the header
+                var length = BitConverter.ToUInt32(glbHeader, 4);
+                if (length >= int.MaxValue)
+                {
+                    // glTF-binary supports up to 2^32 = 4GB, but C# arrays have a 2^31 (2GB) limit.
+                    m_Logger?.Error("glb exceeds 2GB limit.");
+                    return false;
+                }
+                var data = new byte[length];
+                Array.Copy(firstBytes, data, firstBytes.Length);
+                Array.Copy(glbHeader, 0, data, firstBytes.Length, glbHeader.Length);
+                // The amount of bytes we've already read
+                var offset = firstBytes.Length + glbHeader.Length;
+                var pendingBytes = (int)length - offset;
+
+                if (!await stream.ReadToArrayAsync(data, offset, pendingBytes, cancellationToken))
+                {
+                    m_Logger?.Error(LogCode.StreamError, "glb data");
                 }
 
                 return await LoadGltfBinary(data, uri, importSettings, cancellationToken);
             }
-            fs.Close();
+            var reader = new StreamReader(stream);
+            string json;
+            if (stream.CanSeek)
+            {
+                stream.Seek(initialStreamPosition, SeekOrigin.Begin);
+                json = await reader.ReadToEndAsync();
+            }
+            else
+            {
+                // TODO: String concat likely leads to another copy in memory and bad performance.
+                json = Encoding.UTF8.GetString(firstBytes) + await reader.ReadToEndAsync();
+            }
 
-            return await LoadGltfJson(
-#if UNITY_2021_3_OR_NEWER
-                await File.ReadAllTextAsync(localPath,cancellationToken),
-#else
-                File.ReadAllText(localPath),
-#endif
-                uri,
-                importSettings, cancellationToken);
+            reader.Dispose();
+
+            return !cancellationToken.IsCancellationRequested
+                && await LoadGltfJson(json, uri, importSettings, cancellationToken);
         }
 
         /// <summary>
@@ -779,6 +826,18 @@ namespace GLTFast
         }
 
         /// <inheritdoc />
+        public async Task<UnityEngine.Material> GetMaterialAsync(int index)
+        {
+            return await GetMaterialAsync(index, new CancellationToken());
+        }
+
+        /// <inheritdoc />
+        public Task<UnityEngine.Material> GetMaterialAsync(int index, CancellationToken cancellationToken)
+        {
+            return Task.FromResult(GetMaterial(index));
+        }
+
+        /// <inheritdoc />
         public UnityEngine.Material GetDefaultMaterial()
         {
 #if UNITY_EDITOR
@@ -794,6 +853,18 @@ namespace GLTFast
             m_MaterialGenerator.SetLogger(null);
             return material;
 #endif
+        }
+
+        /// <inheritdoc />
+        public async Task<UnityEngine.Material> GetDefaultMaterialAsync()
+        {
+            return await GetDefaultMaterialAsync(new CancellationToken());
+        }
+
+        /// <inheritdoc />
+        public Task<UnityEngine.Material> GetDefaultMaterialAsync(CancellationToken cancellationToken)
+        {
+            return Task.FromResult(GetDefaultMaterial());
         }
 
         /// <summary>
@@ -901,6 +972,41 @@ namespace GLTFast
         }
 
         /// <inheritdoc />
+        public MeshPrimitiveBase GetSourceMeshPrimitive(int meshIndex, int primitiveIndex)
+        {
+            if (Root?.Meshes != null && meshIndex >= 0 && meshIndex < Root.Meshes.Count)
+            {
+                var mesh = Root.Meshes[meshIndex];
+                if (mesh?.Primitives != null && primitiveIndex >= 0 && primitiveIndex < mesh.Primitives.Count)
+                {
+                    return mesh.Primitives[primitiveIndex];
+                }
+            }
+            return null;
+        }
+
+        /// <inheritdoc />
+        public IMaterialsVariantsSlot[] GetMaterialsVariantsSlots(int meshIndex, int meshResultOffset)
+        {
+            var meshResultIndex = m_MeshPrimitiveIndex[meshIndex] + meshResultOffset;
+            Assert.IsTrue(meshResultIndex < m_MeshPrimitiveIndex[meshIndex + 1]);
+
+            List<IMaterialsVariantsSlot> materialSlots = null;
+            var meshResult = m_Primitives[meshResultIndex];
+            foreach (var primitiveIndex in meshResult.primitiveIndices)
+            {
+                var primitive = GetSourceMeshPrimitive(meshIndex, primitiveIndex);
+                if (primitive.Extensions?.KHR_materials_variants?.mappings != null)
+                {
+                    materialSlots ??= new List<IMaterialsVariantsSlot>();
+                    materialSlots.Add(primitive);
+                }
+            }
+
+            return materialSlots?.ToArray();
+        }
+
+        /// <inheritdoc />
         public NodeBase GetSourceNode(int index = 0)
         {
             if (Root?.Nodes != null && index >= 0 && index < Root.Nodes.Count)
@@ -958,6 +1064,15 @@ namespace GLTFast
             }
             var accessor = Root.Accessors[accessorIndex];
             return GetBufferView(accessor.bufferView, accessor.byteOffset, accessor.ByteSize);
+        }
+
+        /// <inheritdoc />
+        public int MaterialsVariantsCount => Root.MaterialsVariantsCount;
+
+        /// <inheritdoc />
+        public string GetMaterialsVariantName(int index)
+        {
+            return Root.GetMaterialsVariantName(index);
         }
 
         async Task<bool> LoadFromUri(Uri url, CancellationToken cancellationToken)