+++ Runtime/Scripts/Schema/FakeSchema/MaterialExtension.cs
@@ -13,6 +13,8 @@ namespace GLTFast.FakeSchema
         public string KHR_materials_transmission;
         public string KHR_materials_clearcoat;
         public string KHR_materials_sheen;
+        public string KHR_materials_ior;
+        public string KHR_materials_specular;
         // ReSharper restore InconsistentNaming
     }
 }