+++ Runtime/Scripts/Schema/MaterialSheen.cs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2023 Unity Technologies and the glTFast authors
 // SPDX-License-Identifier: Apache-2.0
 
+using Unity.Mathematics;
 using UnityEngine;
 
 namespace GLTFast.Schema
@@ -25,12 +26,19 @@ namespace GLTFast.Schema
         /// <summary>
         /// The sheen color in linear space.
         /// </summary>
-        public Color SheenColor =>
-            new Color(
-                sheenColorFactor[0],
-                sheenColorFactor[1],
-                sheenColorFactor[2]
-            );
+        public Color SheenColor
+        {
+            get =>
+                new Color(
+                    sheenColorFactor[0],
+                    sheenColorFactor[1],
+                    sheenColorFactor[2]
+                );
+            set
+            {
+                sheenColorFactor = new[] { value.r, value.g, value.b };
+            }
+        }
 
         /// <summary>
         /// The sheen color texture.
@@ -50,8 +58,29 @@ namespace GLTFast.Schema
         internal void GltfSerialize(JsonWriter writer)
         {
             writer.AddObject();
+            if (sheenColorFactor != null && sheenColorFactor.Length > 2 && (
+                    math.abs(sheenColorFactor[0] - 1f) > Constants.epsilon ||
+                    math.abs(sheenColorFactor[1] - 1f) > Constants.epsilon ||
+                    math.abs(sheenColorFactor[2] - 1f) > Constants.epsilon
+                ))
+            {
+                writer.AddArrayProperty("sheenColorFactor", sheenColorFactor);
+            }
+            if (sheenColorTexture != null)
+            {
+                writer.AddProperty("sheenColorTexture");
+                sheenColorTexture.GltfSerialize(writer);
+            }
+            if (sheenRoughnessFactor > 0)
+            {
+                writer.AddProperty("sheenRoughnessFactor", sheenRoughnessFactor);
+            }
+            if (sheenRoughnessTexture != null)
+            {
+                writer.AddProperty("sheenRoughnessTexture");
+                sheenRoughnessTexture.GltfSerialize(writer);
+            }
             writer.Close();
-            throw new System.NotImplementedException($"GltfSerialize missing on {GetType()}");
         }
     }
 }