+++ Runtime/Scripts/Schema/Root.cs
@@ -567,6 +567,16 @@ namespace GLTFast.Schema
                     {
                         mat.Extensions.KHR_materials_sheen = null;
                     }
+
+                    if (fake.KHR_materials_ior == null)
+                    {
+                        mat.Extensions.KHR_materials_ior = null;
+                    }
+
+                    if (fake.KHR_materials_specular == null)
+                    {
+                        mat.Extensions.KHR_materials_specular = null;
+                    }
                 }
             }
 
@@ -627,6 +637,29 @@ namespace GLTFast.Schema
             {
                 UnsetExtensions();
             }
+        }
+
+        /// <summary>
+        /// Number of materials variants.
+        /// </summary>
+        /// <seealso href="https://github.com/KhronosGroup/glTF/tree/main/extensions/2.0/Khronos/KHR_materials_variants"/>
+        public int MaterialsVariantsCount => Extensions?.KHR_materials_variants?.variants?.Count ?? 0;
+
+        /// <summary>
+        /// Gets the name of a specific materials variant.
+        /// </summary>
+        /// <param name="index">Materials variant index.</param>
+        /// <returns>Name of a materials variant.</returns>
+        /// <seealso href="https://github.com/KhronosGroup/glTF/tree/main/extensions/2.0/Khronos/KHR_materials_variants"/>
+        public string GetMaterialsVariantName(int index)
+        {
+            var variants = Extensions?.KHR_materials_variants?.variants;
+            if (variants != null && index >= 0 && index < variants.Count)
+            {
+                return variants[index].name;
+            }
+
+            return null;
         }
     }
 }